backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
distlib==0.3.9
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docker==7.1.0
filelock==3.18.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
nodeenv==1.9.1
openai==1.82.0
packaging==25.0
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
